Available functions

CYAxiverse.filestructure

CYAxiverse.filestructure — Module

CYAxiverse.filestructure

This module sets up the structure of the database, identifying where to locate data / plot files etc

CYAxiverse.filestructure.Kfile — Function

Kfile(h11,tri,cy)

Loads Kähler metric specified by geometry index.

Warning

Deprecated

CYAxiverse.filestructure.Lfile — Function

Lfile(h11,tri,cy)

Loads instanton energy scales specified by geometry index.

Warning

Deprecated

CYAxiverse.filestructure.Qfile — Function

Qfile(h11,tri,cy)

Loads instanton charge matrix specified by geometry index.

Warning

Deprecated

CYAxiverse.filestructure.cyax_file — Function

cyax_file(h11,tri,cy)

Path to data file – will contain all data that relates to geometry index.

CYAxiverse.filestructure.data_dir — Method

data_dir()

Creates/reads data directory

CYAxiverse.filestructure.geom_dir — Function

geom_dir(h11,tri,cy)

Defines file directories for data specified by geometry index.

CYAxiverse.filestructure.h11lst — Function

h11lst(min,max)

Loads geometry indices between $h^{1,1} \in (\mathrm{min},\mathrm{max}]$

CYAxiverse.filestructure.h11lst — Method

h11lst(h11list::Vector)

TBW

CYAxiverse.filestructure.isgeometry — Method

isgeometry(h11, tri, cy)

Check if geometric quantities have been computed

CYAxiverse.filestructure.localARGS — Method

localARGS()

Load key for data dir – key should be in ol_DB

CYAxiverse.filestructure.log_dir — Method

log_dir()

Creates/reads log directory

CYAxiverse.filestructure.logcreate — Method

logcreate(l)

Creates logfile

CYAxiverse.filestructure.logfile — Method

logfile()

Returns path of logfile in format data_dir()/logs/YYYY:MM:DD:T00:00:00.000log.out

CYAxiverse.filestructure.minfile — Function

minfile(h11,tri,cy)

Path to file containing minimization data.

CYAxiverse.filestructure.np_path — Method

np_path()

Saves list of data paths and matrix of [h11; tri; cy] in h5 file paths_cy.h5

CYAxiverse.filestructure.np_path_generate — Method

np_path_generate(h11)

Walks through data_dir() and returns list of data paths and matrix of [h11; tri; cy] – at specific h11. Saves in h5 file paths_cy.h5

CYAxiverse.filestructure.np_path_generate — Method

np_path_generate()

Walks through data_dir() and returns list of data paths and matrix of [h11; tri; cy]. Saves in h5 file paths_cy.h5

CYAxiverse.filestructure.ol_DB — Method

ol_DB(args)

Define dict of directories for data read/write

CYAxiverse.filestructure.paths_cy — Method

paths_cy()

Loads / generates paths_cy.h5 which contains the explicit locations and also [h11; tri; cy] indices of the geometries already saved.

CYAxiverse.filestructure.plots_dir — Method

plots_dir()

Creates/reads a directory for plots

CYAxiverse.filestructure.present_dir — Method

present_dir()

Returns the present data directory using localARGS

CYAxiverse.generate

CYAxiverse.generate — Module

CYAxiverse.generate

This is where most of the functions are defined.

CYAxiverse.generate.LQtildebar — Method

LQtildebar(L,Q; threshold)

Compute the linearly independent leading instantons that generate the axion potential, including any subleading instantons that are within threshold of their basis instanton.

#Examples

julia> h11,tri,cy = 8, 108, 1;
julia> pot_data = CYAxiverse.read.potential(h11,tri,cy);
julia> vacua_data = CYAxiverse.generate.LQtildebar(pot_data["L"],pot_data["Q"]; threshold=0.5)
Dict{String, Matrix{Float64}}(
"Ltilde" => 2×9 Matrix{Float64}:
   1.0       1.0       1.0       1.0     …     1.0       1.0       1.0       1.0
 -46.5338  -49.2362  -57.3519  -92.6082     -136.097  -163.303  -179.634  -163.303

"Lbar" => 2×12 Matrix{Float64}:
   1.0      1.0        1.0       1.0    …    -1.0      -1.0       1.0       1.0
 -90.414  -98.5299  -101.232  -133.787     -177.681  -177.681  -179.978  -179.978

"Qbar" => 8×12 Matrix{Float64}:
  0.0   0.0   0.0   2.0   2.0   2.0  1.0   0.0   0.0   0.0   0.0   0.0
  0.0   0.0   0.0   0.0   0.0   0.0  0.0   0.0   0.0   0.0   0.0   0.0
  0.0   0.0   0.0  -1.0  -1.0  -1.0  0.0  -1.0   1.0   0.0   1.0   0.0
  0.0   0.0   0.0  -1.0  -1.0  -1.0  0.0   1.0   0.0   1.0   0.0   1.0
  0.0   0.0   0.0   0.0   0.0   0.0  0.0   1.0   0.0   0.0   0.0   0.0
  1.0   1.0   0.0   0.0   1.0   1.0  0.0   0.0  -1.0  -1.0   0.0   0.0
 -1.0   0.0   1.0   1.0   0.0   1.0  0.0   0.0   0.0   0.0  -1.0  -1.0
  0.0  -1.0  -1.0  -1.0  -1.0  -2.0  0.0   0.0   0.0   0.0   0.0   0.0

"α" => 12×8 Matrix{Float64}:
  1.0  -1.0   0.0  0.0   0.0  0.0  0.0  0.0
  1.0   0.0  -1.0  0.0   0.0  0.0  0.0  0.0
  0.0   1.0  -1.0  0.0   0.0  0.0  0.0  0.0
 -1.0   0.0   0.0  1.0   0.0  0.0  0.0  0.0
  0.0  -1.0   0.0  1.0   0.0  0.0  0.0  0.0
  0.0   0.0  -1.0  1.0   0.0  0.0  0.0  0.0
  0.0   0.0   0.0  0.0   0.0  0.0  0.0  0.0
  0.0   0.0   0.0  0.0  -1.0  1.0  1.0  0.0
 -1.0   0.0   0.0  0.0   1.0  0.0  0.0  0.0
 -1.0   0.0   0.0  0.0   0.0  1.0  0.0  0.0
  0.0  -1.0   0.0  0.0   1.0  0.0  0.0  0.0
  0.0  -1.0   0.0  0.0   0.0  1.0  0.0  0.0

"Qtilde" => 8×9 Matrix{Float64}:
 0.0  0.0  0.0   2.0  0.0  0.0  0.0  0.0   0.0
 0.0  0.0  0.0   0.0  0.0  0.0  0.0  1.0   0.0
 0.0  0.0  0.0  -1.0  1.0  0.0  0.0  0.0  -1.0
 0.0  0.0  0.0  -1.0  0.0  1.0  0.0  0.0   1.0
 0.0  0.0  0.0   0.0  0.0  0.0  1.0  0.0   1.0
 1.0  0.0  0.0   1.0  0.0  0.0  0.0  0.0   0.0
 0.0  1.0  0.0   1.0  0.0  0.0  0.0  0.0   0.0
 0.0  0.0  1.0  -1.0  0.0  0.0  0.0  0.0   0.0
 )

CYAxiverse.generate.constants — Method

constants()

Loads constants:

Reduced Planck Mass = 2.435 × 10^18

Hubble = 2.13 × 0.7 × 10^-33

log2pi = log10(2π)

as Dict{String,ArbFloat}

#Examples

julia> const_data = CYAxiverse.generate.constants()
Dict{String, ArbNumerics.ArbFloat{128}} with 3 entries:
  "MPlanck" => 2435000000000000000.0
  "log2π"   => 0.7981798683581150521959557408991
  "Hubble"  => 1.490999999999999999287243983194e-33

CYAxiverse.generate.gauss_diff — Method

gauss_diff(z)

Computes the difference of 2 numbers in (natural) log-space using the definition here.

#Examples

julia> CYAxiverse.generate.gauss_diff(10.)
9.99995459903963

julia> CYAxiverse.generate.gauss_diff(1000.)
1000.0

CYAxiverse.generate.gauss_log_split — Method

gauss_log_split(sign, log)

Algorithm to compute Gaussian logarithms, as detailed here.

#Examples

julia> CYAxiverse.generate.gauss_diff(10.)
9.99995459903963

julia> CYAxiverse.generate.gauss_diff(1000.)
1000.0

CYAxiverse.generate.gauss_sum — Method

gauss_sum(z)

Computes the addition of 2 numbers in (natural) log-space using the definition here.

#Examples

julia> CYAxiverse.generate.gauss_sum(10.)
10.000045398899218

julia> CYAxiverse.generate.gauss_sum(1000.)
1000.0

CYAxiverse.generate.hp_spectrum — Method

hp_spectrum(K,L,Q; prec=5_000)

Uses potential data generated by CYTools (or randomly generated) to compute axion spectra – masses, quartic couplings and decay constants – to high precision.

#Examples

julia> pot_data = CYAxiverse.read.potential(4,10,1)
julia> hp_spec_data = CYAxiverse.generate.hp_spectrum(pot_data["K"], pot_data["L"], pot_data["Q"])
Dict{String, Any} with 12 entries:
    "msign" => []
    "m" => []
    "fK" => []
    "fpert" => []
    "λselfsign" => []
    "λself" => []
    "λ31_i" => []
    "λ31sign" => []
    "λ31" => []
    "λ22_i" => []
    "λ22sign" => []
    "λ22" => []

CYAxiverse.generate.hp_spectrum_save — Function

hp_spectrum_save(h11,tri,cy)

CYAxiverse.generate.orth_basis — Method

orth_basis(Q)

Takes a set of vectors (columns of Q) and constructs an orthonormal basis

CYAxiverse.generate.pq_spectrum — Method

pq_spectrum(K,L,Q)

Uses a modified version of the algorithm outlined in the PQ Axiverse paper (Appendix A) to compute the masses and decay constants.

Note

The off-diagonal elements of the quartic self-interaction tensor are not yet included in this computation

CYAxiverse.generate.pseudo_K — Function

pseudo_K(h11,tri,cy=1)

Randomly generates an h11 × h11 Hermitian matrix with positive definite eigenvalues.

#Examples

julia> K = CYAxiverse.generate.pseudo_K(4,10,1)
4×4 Hermitian{Float64, Matrix{Float64}}:
 2.64578  2.61012  0.91203  2.27339
 2.61012  3.89684  2.22451  1.93356
 0.91203  2.22451  2.94717  1.58126
 2.27339  1.93356  1.58126  4.85208

julia> eigen(K).values
4-element Vector{Float64}:
 0.17629073145135896
 1.8632009739875723
 2.7425362219513487
 9.559840749713599

CYAxiverse.generate.pseudo_L — Function

pseudo_L(h11,tri,cy=1;log=true)

Randomly generates a h11+4+C(h11+4,2)-length hierarchical list of instanton scales, similar to those found in the KS Axiverse. Option for (sign,log10) or full precision.

#Examples

julia> CYAxiverse.generate.pseudo_L(4,10)
36×2 Matrix{Float64}:
  1.0     0.0
  1.0    -4.0
  1.0    -8.0
  1.0   -12.0
  1.0   -16.0
  1.0   -20.0
  1.0   -24.0
  1.0   -28.0
 -1.0   -29.4916
  1.0   -33.8515
  ⋮
  1.0  -133.665
 -1.0  -138.951

julia> CYAxiverse.generate.pseudo_L(4,10,log=false)
36-element Vector{ArbNumerics.ArbFloat}:
  1.0
  0.0001
  1.0e-8
  1.0e-12
  1.0e-16
  1.0e-20
  1.0e-24
  1.0e-28
 -1.462574279422558833057690597964e-31
 -2.381595397961591074099629406235e-34
  ⋮
  3.796809523142314798130344022481e-134
 -3.173000613781491329619833894919e-138

CYAxiverse.generate.pseudo_Q — Function

pseudo_Q(h11,tri,cy=1)

Randomly generates an instanton charge matrix that takes the same form as those found in the KS Axiverse, namely I(h11) with 4 randomly filled rows and the cross-terms, i.e. an h11+4+C(h11+4,2) × h11 integer matrix.

#Examples

julia> CYAxiverse.generate.pseudo_Q(4,10,1)
36×4 Matrix{Int64}:
  1   0   0   0
  0   1   0   0
  0   0   1   0
  0   0   0   1
  1   4  -3   5
 -5  -4  -2   4
  4   5   3  -2
 -5   2  -3  -3
  ⋮
 -9  -9  -5   6
  0  -6   1   7
  9   3   6   1

CYAxiverse.generate.vacua — Method

vacua(L,Q; threshold)

Compute the number of vacua given an instanton charge matrix Q and 2-column matrix of instanton scales L (in the form [sign; exponent]) and a threshold for:

$\frac{\Lambda_a}{|\Lambda_j|}$

i.e. is the instanton contribution large enough to affect the minima.

For small systems (Nax<=50) the algorithm computes the ratio of volumes of the fundamental domain of the leading potential and the full potential.

For larger systems, the algorithm only computes the volume of the fundamental domain of the leading potential.

#Examples

julia> using CYAxiverse
julia> h11,tri,cy = 10,20,1;
julia> pot_data = CYAxiverse.read.potential(h11,tri,cy);
julia> vacua_data = CYAxiverse.generate.vacua(pot_data["L"],pot_data["Q"])
Dict{String, Any} with 3 entries:
  "θ∥"     => Rational[1//1 0//1 … 0//1 0//1; 0//1 1//1 … 0//1 0//1; … ; 0//1 0//1 … 1//1 0//1; 0//1 0//1 … 0//1 1//1]
  "vacua"  => 3
  "Qtilde" => [0 0 … 1 0; 0 0 … 0 0; … ; 1 1 … 0 0; 0 0 … 0 0]

CYAxiverse.generate.vacua_MK — Method

vacua_MK(L,Q; threshold=1e-2)

Uses the projection method of PQ Axiverse paper (Appendix A) on $\mathcal{Q}$ to compute the locations of vacua.

Note

Finding the lattice of minima when numerical minimisation is required has not yet been implemented.

CYAxiverse.generate.vacua_MK — Method

vacua_MK(L,Q; threshold=1e-2)

Uses the projection method of PQ Axiverse paper (Appendix A) on $\mathcal{Q}$ to compute the locations of vacua.

Note

Finding the lattice of minima when numerical minimisation is required has not yet been implemented.

CYAxiverse.generate.vacua_TB — Method

vacua_TB(h11,tri,cy)

Compute the number of vacua given a geometry from the KS database.

For small systems (Nax<=50) the algorithm computes the ratio of volumes of the fundamental domain of the leading potential and the full potential.

For larger systems, the algorithm only computes the volume of the fundamental domain of the leading potential. #Examples

julia> using CYAxiverse
julia> h11,tri,cy = 10,20,1;
julia> vacua_data = CYAxiverse.generate.vacua_TB(h11,tri,cy)
Dict{String, Any} with 3 entries:
  "θ∥"     => Rational[1//1 0//1 … 0//1 0//1; 0//1 1//1 … 0//1 0//1; … ; 0//1 0//1 … 1//1 0//1; 0//1 0//1 … 0//1 1//1]
  "vacua"  => 11552.0
  "Qtilde" => [0 0 … 0 1; 0 0 … 0 0; … ; 1 1 … -1 -1; 0 0 … 0 0]

CYAxiverse.generate.vacua_TB — Method

vacua_TB(L,Q)

Compute the number of vacua given an instanton charge matrix Q and 2-column matrix of instanton scales L (in the form [sign; exponent])

For small systems (Nax<=50) the algorithm computes the ratio of volumes of the fundamental domain of the leading potential and the full potential.

For larger systems, the algorithm only computes the volume of the fundamental domain of the leading potential. #Examples

julia> using CYAxiverse
julia> h11,tri,cy = 10,20,1;
julia> pot_data = CYAxiverse.read.potential(h11,tri,cy);
julia> vacua_data = CYAxiverse.generate.vacua_TB(pot_data["L"],pot_data["Q"])
Dict{String, Any} with 3 entries:
  "θ∥"     => Rational[1//1 0//1 … 0//1 0//1; 0//1 1//1 … 0//1 0//1; … ; 0//1 0//1 … 1//1 0//1; 0//1 0//1 … 0//1 1//1]
  "vacua"  => 11552.0
  "Qtilde" => [0 0 … 0 1; 0 0 … 0 0; … ; 1 1 … -1 -1; 0 0 … 0 0]

CYAxiverse.generate.vacua_full — Method

vacua_MK(L,Q; threshold=1e-2)

New implementation of MK's algorithm – testing!

CYAxiverse.generate.vacua_id — Method

vacua_id(L, Q; threshold)

CYAxiverse.generate.vacua_id_basis — Method

vacua_id_basis(L,Q; threshold)

Compute the number of vacua given an instanton charge matrix Q and 2-column matrix of instanton scales L (in the form [sign; exponent]) and a threshold for:

$\frac{\Lambda_a}{|\Lambda_j|}$

i.e. is the instanton contribution large enough to affect the minima. This function uses JLM's method outlined in [TO APPEAR].

#Examples

julia> using CYAxiverse
julia> h11,tri,cy = 10,20,1;
julia> pot_data = CYAxiverse.read.potential(h11,tri,cy);
julia> vacua_data = CYAxiverse.generate.vacua_id_basis(pot_data["L"],pot_data["Q"]; threshold=0.01)
Dict{String, Any} with 3 entries:
  "θ∥"     => Rational[1//1 0//1 … 0//1 0//1; 0//1 1//1 … 0//1 0//1; … ; 0//1 0//1 … 1//1 0//1; 0//1 0//1 … 0//1 1//1]
  "vacua"  => 11552.0
  "Qtilde" => [0 0 … 0 1; 0 0 … 0 0; … ; 1 1 … -1 -1; 0 0 … 0 0]

CYAxiverse.minimizer

CYAxiverse.minimizer — Module

CYAxiverse.minimizer

Some minimization / optimization routines to find vacua and other such explorations.

CYAxiverse.minimizer.grad_std — Method

grad_std(h11::Int,tri::Int,cy::Int,LV::Vector,QV::Matrix)

TBW

CYAxiverse.minimizer.grad_std — Method

grad_std(LV::Vector,QV::Matrix)

TBW

CYAxiverse.minimizer.minima_lattice — Method

minima_lattice(v::Matrix{Float64})

TBW

CYAxiverse.minimizer.minimize — Method

minimize(LV::Vector,QV::Matrix,x0::Vector)

TBW

CYAxiverse.minimizer.subspace_minimize — Method

subspace_minimize(L::Vector, Q::Matrix; runs=10_000)

Minimizes the subspace with runs iterations

CYAxiverse.read

CYAxiverse.read — Module

CYAxiverse.read

Functions that access the database.

CYAxiverse.cytools_wrapper

Missing docstring.

Missing docstring for add_functions/CYAxiverse.cytools_wrapper. Check Documenter's build log for details.

Modules = [CYAxiverse.cytools_wrapper]
Pages = ["add_functions/cytools_wrapper.jl"]